backend/requirements.txt
anyio==4.11.0
APScheduler==3.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.7
django-apscheduler==0.7.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
mutagen==1.47.0
pillow==12.0.0
psycopg==3.2.11
psycopg-binary==3.2.11
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.3
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
# uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
whitenoise==6.11.0
